// src/enums/StockTrading.tsx
export const enum StockTrading {
    US_Buy = "VTTT1002U",   // 미국 매수 주문
    US_Sell = "VTTT1001U",  // 미국 매도 주문
    JP_Buy = "VTTS0308U",   // 일본 매수 주문
    JP_Sell = "VTTS0307U",  // 일본 매도 주문
    CN_SH_Buy = "VTTS0202U",// 상해 매수 주문
    CN_SH_Sell = "VTTS1005U",// 상해 매도 주문
    HK_Buy = "VTTS1002U",   // 홍콩 매수 주문
    HK_Sell = "VTTS1001U",  // 홍콩 매도 주문
    CN_SZ_Buy = "VTTS0305U",// 심천 매수 주문
    CN_SZ_Sell = "VTTS0304U",// 심천 매도 주문
    VN_Buy = "VTTS0311U",   // 베트남 매수 주문
    VN_Sell = "VTTS0310U"   // 베트남 매도 주문
}